Additional scheduler features and bug tracker fixes (#127)
* Expand and collapse changes

* License added in files and launcher card order

* status logo dark theme fixes

* Dark mode color fixes

* State color change

* CSS fixes in execution history page

* version updated to 0.1.73

* Create and list dag flow added and logo updates

* antd package added

* dag template start cluster changes

* Header name changes and disable HTML checkbox

* Create job scheduler UI fixes

* Download added inside actions column

* Title updated in execution history

* By default selection changes

* calendar date border radius removed

* dag run list empty condition handled

* date wise filter moved to FE logic

* Default highlight final dag runs changes

* Select current date by default changes

* expand less and more added

* By default open first logs

* GCS innerHTML to 'textContent' change

* Orange color code Icon added as badge

* template changes

* dark theme changes

* tags changes

* Fix -Create Scheduler:Header and back button, List:removed default for composer List

* create scheduler : back and cancel fix

* create scheduler fix default timezone UTC

* Job name field fix

* removing local composer code

* removing conditions

* compose dropdown fix

* removing local environment

* passiing parameters changes

* Create Scheduler Job name and email Fix Listing toast added for empty composer list

* build test

* build check

* Build check

* parameter changes

* Execution History Heading and Back button fixed

* Listing page- delete confirmation pop up

* create page :email fix, listing :no composer list fix

* create page composer list passing to list page

* composer passing to list page

* composer fetching

* empty parameter handling

* email success addition

* Create scheduler payload fix according to cluster and serverless

* serverless changes

* name change

* papermill path changes

* List page scroll bar fix

* timezone fixed

* code cleanup and dag template changes

* output format fix

* removing print statements

* Revert "removing print statements"

This reverts commit 9e1f874c12fd5b23c3601625e892a25708679451.

* removing print statements

* Adding license

* timezone and sorting fixed

* back button and header spacing fixed

* build-check

* scroll bar fix

* review comments and path change

* indentation fix

* adding pydantic dependency

* pydantic version change

* adding license

* css fix

* reverted orange color icon for date

* indendation fix

* No rows to display fixed

* future date fix

* code clean up

* adding logs

* beta command changes

* Edit Dag backend changes

* Edit dag API integration and FE changes

* adding logs param

* Edit job values set using edit dag response

* payload dummy values modified

* UTC added as dummy value

* Edit all values handled with fixed dummy entry

* parameter test

* passing parameters

* stop cluster param fix

* polling implemented

* Execution history selected row breakage fix , updated toast for delete

* parameter fix

* email fix

* Check boxes and parameter detail added

* code cleanup

* Launcher flow for Edit dag fixes

* Loader added for Edit api call

* Serverless name  state updation

* Serverless data update added for edit

* Error handling in create dag api

* Serverless mode Edit dag changes

* handling input file and file read changes

* Bug tracker fixes - ID 108, 109

* calendar date display in circle border

* Edit notebook from GCS changes

* edit notebook enable condition changed

* code cleanup

* input file handling for edit

* JobName unique implemented

* execution_date to start_date change

* Edit notebook and Timezone fix

1. Edit notebook added in listing page
2. Timezone default value added

* adding time_zone to dag

* removing additional char

* fetching timeZone

* handling time_zone for old dags

* schedule value name fix

* timezone spelling

* title text change for edit dag

* timezone to time_zone fix

* Edit notebook icon and parameters clone fix

* timezone empty condition handled

* Delete dag BE changes

* Create and Update text changes

* Run schedule toggle fix

* Edit mode Fixes, icon removed and notebook edit fixes

* Jobname and composer disable on edit mode

* Edit mode back and cancel action change

* cancel disable on creating/updating

* delete pop-up fix

* Update toast changed

* Client bug tracker fixes - 123, 125

* client bug tracker 126-show years from 2024

* 4.1.0 version fixes - 105, 106

* container image and metastore template changes

* Client bug fix id - 124

* adding catchup in template

* catchup fix

* Job name unique, action button sequence on listing,delete dag toast fixed

* DPMS loading on other screens fix

* removing addtional field

* dag controller classname changes, adding tornado authenticator

* adding catchup to dag

* removing import controller

* review comments fixes

* Unused commented code cleanup

* Code review comments change

* Sonar Fixes

* Sonar scan UI fixes

* Sonar scan FE changes

* Dag run pagination and offset changes

* dag run height, width fixes

* Import error popup

* Scheduler new icons updated from figma

* Trigger Dag API BE

* Dag run Offset added changes

* Session list api - empty handled

* reverted polling for import error

* Download output notebook BE

* output notebook changes

* Download output notebook FE changes

* Download output BE changes

* trigger the dag implemented

* Dag run state clicked and code cleanup

* Trigger and play icon updated

* gcloud command downlaod changes

* path changes

* status code change

* Revert "status code change"

This reverts commit c8b9aee2b395a1ab5f594681987432c810e5172a.

* disable download based on status

* status change and path change

* dagRunId encoded for api call

* download output code cleanup

* download disable for failed state

* dag run list filter added for null

* Title changed for output notebook

* Download output loader added in FE

* import error and trigger dag

Import error api polling 30 sec
delete for import error
changed name to schedule error
trigger the dag disable on pause

* from page  change

* status change

* Delete import error fixes

* Calendar color code changes in execution history

* Code review fixes

* code review fixes

* pagination reset on date change

* MUi Button change

* schedule Mode changes

* button change

* Daglist Interface

* code cleanup

* Feedback UI changes

* Code review changes BE

1. changing structure of checking credentials
2. Code formatting

* multiple toast fix

* unused code cleanup

* BE code review comment fixes for new files

* FE code review changes:

1. License file re generated
2. page size increased

* Code review fixes BE

* Code cleanup

* label properties default value bugfix

* task name change and format fix

* Update wrapper_papermill.py

Removed the unwanted lines of code

* review comment fix

* code review comments fix

* last code review comment

* Code review changes

* code cleanup

---------

+++ dataproc_jupyter_plugin/services/dagListService.py
@@ -74,7 +74,7 @@ class DagListService:
 
 
 class DagDeleteService:
-    def delete_job(self, credentials, composer_name, dag_id, log):
+    def delete_job(self, credentials, composer_name, dag_id, from_page, log):
         airflow_uri, bucket = DagListService.get_airflow_uri(
             self, composer_name, credentials, log
         )
@@ -86,8 +86,9 @@ class DagDeleteService:
                     "Content-Type": CONTENT_TYPE,
                     "Authorization": f"Bearer {access_token}",
                 }
-                response = requests.delete(api_endpoint, headers=headers)
-                log.info(response)
+                if from_page == None:
+                    response = requests.delete(api_endpoint, headers=headers)
+                    log.info(response)
                 cmd = f"gsutil rm gs://{bucket}/dags/dag_{dag_id}.py"
                 process = subprocess.Popen(
                     cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True
